Secure Boot clarifications for Collector
Collector just became smarter about UEFI Secure Boot [1], and now will
switch to eBPF when needed to make the user experience smoother.
Represent this in the documentation, including clarification that
Collector will not work with kernel modules on UEFI when Secure Boot is
enabled (not on any system booted in UEFI mode).

[1] https://github.com/stackrox/collector/pull/553